% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/for_server_files.R
\name{breakout_server_code}
\alias{breakout_server_code}
\title{Parse server file for assignments & inputst}
\usage{
breakout_server_code(file)
}
\arguments{
\item{file}{file to parse}
}
\description{
Parse server file for assignments & inputst
}
\examples{
\dontrun{
breakout_server_code(file = "inst/shiny/server.R") \%>\% 
  eval(envir = .GlobalEnv)
}
}
