% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/for_server_files.R
\name{char_between}
\alias{char_between}
\title{Find text between brackets or parentheses}
\usage{
char_between(text, pattern = c("c", "p"))
}
\arguments{
\item{text}{the text to scan}

\item{pattern}{look between "c" curly braces or "p" parenthese}
}
\description{
Find text between brackets or parentheses
}
\examples{
"here is text (between two parentheses), as an example" \%>\% 
  char_between("p")
  
"here is text {between two curly braces}, as an example" \%>\% 
  char_between("c")
  
}
