% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runapp_shinyapp_innards.R
\name{deparse_server}
\alias{deparse_server}
\title{Returns asignments only from expressions}
\usage{
deparse_server(x)
}
\arguments{
\item{x}{expression from code stored as text}
}
\description{
Returns asignments only from expressions
}
