% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runapp_shinyapp_innards.R
\name{extract_from_app_fn}
\alias{extract_from_app_fn}
\title{Returns server code from shinyApp or runApp}
\usage{
extract_from_app_fn(text)
}
\arguments{
\item{text}{code stored as text}
}
\description{
Returns server code from shinyApp or runApp
}
