% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{find_input_code}
\alias{find_input_code}
\title{Look for input <- demo}
\usage{
find_input_code(file)
}
\arguments{
\item{file}{to evaluate}
}
\description{
Look for input <- demo
}
\examples{
\dontrun{
find_input_code("inst/shiny/server.R")
find_input_code("inst/Rmd/flexdashboard_demo.Rmd")
}
}
