% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runapp_shinyapp_innards.R
\name{guts}
\alias{guts}
\title{Returns what was called inside of a function stored as text}
\usage{
guts(x, fn)
}
\arguments{
\item{x}{the code stored as text}

\item{fn}{the function to sub out}
}
\description{
Returns what was called inside of a function stored as text
}
\examples{
x <- "mean(list(1:10, 10:30))"
guts(x, "mean")
guts(x, "mean") \%>\% guts("list")

}
