% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{input_usage}
\alias{input_usage}
\title{Look for input usage}
\usage{
input_usage(file)
}
\arguments{
\item{file}{file to evaluate}
}
\description{
Prints a statement about the inputs that are either listed or missing
}
\examples{
\dontrun{
input_usage(file = "inst/shiny/server.R")
input_usage(file = "inst/Rmd/flexdashboard_demo.Rmd")
}

}
