% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runapp_shinyapp_innards.R
\name{inside_runapp}
\alias{inside_runapp}
\alias{inside_shinyapp}
\title{Pulls the calls out of runApp and shinyApp}
\usage{
inside_runapp(x)

inside_shinyapp(x)
}
\arguments{
\item{x}{expression containing runApp(...)}
}
\description{
Pulls the calls out of runApp and shinyApp
}
