% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equation_writing.R
\name{make_colors}
\alias{make_colors}
\title{make_colors}
\usage{
make_colors(components)
}
\arguments{
\item{components}{n/a}
}
\description{
function to make color list to match components
}
\keyword{internal}
