% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equation_writing.R
\name{make_equation}
\alias{make_equation}
\title{make_equation}
\usage{
make_equation(parameters, print_colours = TRUE)
}
\arguments{
\item{parameters}{parameters to make the equation}

\item{print_colours}{do you wantt to print colors?}
}
\description{
make equation
}
\keyword{internal}
