% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equation_writing.R
\name{manyToggle}
\alias{manyToggle}
\title{manyToggle}
\usage{
manyToggle(show = NULL, hide = NULL)
}
\arguments{
\item{show}{n/a}

\item{hide}{n/a}
}
\description{
multiple toggles
}
\keyword{internal}
