% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equation_writing.R
\name{order_components}
\alias{order_components}
\title{order_components}
\usage{
order_components(components)
}
\arguments{
\item{components}{n/a}
}
\description{
function to order component names
}
\keyword{internal}
