% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/for_server_files.R
\name{parse_server_file}
\alias{parse_server_file}
\title{Parse server file for assignments & inputst}
\usage{
parse_server_file(file)
}
\arguments{
\item{file}{file to parse}
}
\description{
Parse server file for assignments & inputst
}
\examples{
\dontrun{
parse_server_file(file = "inst/shiny/server.R") \%>\% 
  eval(envir = .GlobalEnv)
}
}
