% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{remove_objects}
\alias{remove_objects}
\title{Clear all objects in environment}
\usage{
remove_objects(keep = NULL)
}
\arguments{
\item{keep}{A regular expression of objects in environment to keep}
}
\description{
Clear all objects in environment
}
\examples{
\dontrun{
df <- iris
df2 <- iris
my_df <- iris
remove_objects(keep = "^df")
}
}
