% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{shinySim}
\alias{shinySim}
\title{shinySim}
\usage{
shinySim(data.struc = NULL)
}
\value{
A simulated dataframe
}
\description{
shiny interface to the squidSim package.
}
\examples{
\donttest{
data_test <- squidSim::make_structure("sex(2)/individual(10)",repeat_obs=2,level_names=list(sex=c("F","M")))
shinySim(data.struc = data_test)
}
}
