% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulated_variances.R
\name{simVar}
\alias{simVar}
\title{simVar}
\usage{
simVar(parameters, data_structure)
}
\arguments{
\item{parameters}{parameter list}

\item{data_structure}{data structure}
}
\value{
A list with means and variance at each level
}
\description{
Calculate simulated mean and variance in response variable(s)
}
\details{
Calculates the simulated variance from the simulation parameters, as well as breaking down the source of this variance into the hierarchical levels given in the parameter list, and into individual predictors. This function has several limitations.
\enumerate{
\item Doesn't work when 'model' is specified in simulate_population()
\item Assumes grouping factors in the data_structure are balanced - unbalanced designs can change the observed variance
\item Will be inaccurate with transformed variables (i.e. if functions are specified)
\item Doesn't deal well with three way interactions and over, unless they are the same variable (i.e. polynomials)
\item Doesn't account for covariance between interaction terms, e.g. if there is are two interaction terms rain:temp and wind:temp, they will covary, but this additional variance in the response is not calculated
}
}
\examples{
\dontrun{

simulated_variance(squid_data)
}
}
\author{
Joel Pick - joel.l.pick@gmail.com
}
\keyword{internal}
