% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{validate_inputs}
\alias{validate_inputs}
\title{Validate demo input statement}
\usage{
validate_inputs(file, output)
}
\arguments{
\item{output}{}
}
\description{
Prints a statement about the inputs that are either listed or missing
}
