% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{which_file}
\alias{which_file}
\title{Select file to use}
\usage{
which_file(file)
}
\arguments{
\item{file}{path to file.}
}
\description{
If the file is not specified, a menu will appear asking the user if they want to use the active source file loaded in RStudio or if they want to select the file (opens a new window).
}
