% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/equation_writing.R
\name{write_code_part}
\alias{write_code_part}
\title{write_code_part}
\usage{
write_code_part(x, print_colours)
}
\arguments{
\item{x}{n/a}

\item{print_colors}{n/a}
}
\description{
writes code
}
\keyword{internal}
