% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{validate_inputs}
\alias{validate_inputs}
\title{Validate demo input statement}
\usage{
validate_inputs(file)
}
\arguments{
\item{file}{file to evaluate}
}
\description{
Prints a statement about the inputs that are either listed or missing
}
\examples{
\dontrun{
validate_inputs("inst/Rmd/test_dashboard_missing_inputs.Rmd")
validate_inputs("inst/Rmd/test_dashboard_no_inputs.Rmd")
validate_inputs("inst/shiny/server.R")
}
}
